Removing domain on liability_account
As e.g. in Germany money received for a gift card (not targeted at a special product) isn't considered to be a revenue, but just an exchange of money, the domain on the liability account is too strict. In Germany it would have to be of kind 'others', but other countries could require different behavior. So I propose to remove the domain from the account.